+++ adminer/type.inc.php
@@ -39,7 +39,8 @@ function add_enum_values(string $type, string $old, string $new): ?array {
 $TYPE = $_GET["type"];
 $row = $_POST;
 // types(true) is the list used for the links to this page in table structure
-$type = ($TYPE != "" ? type_definition(+array_search($TYPE, types(true))) : array());
+$type_id = ($TYPE != "" ? array_search($TYPE, types(true)) : 0);
+$type = ($type_id ? type_definition(+$type_id) : array()); // 0 - the type is being created, false - it doesn't exist
 $object = ($type["kind"] == 'd' ? "DOMAIN" : "TYPE"); // domains are created, altered and dropped as DOMAIN
 
 if ($_POST && !$error) {
@@ -87,7 +88,13 @@ if ($_POST && !$error) {
 	}
 }
 
-page_header($TYPE != "" ? lang('Alter type') . ": " . h($TYPE) : lang('Create type'), $error);
+page_header(
+	($TYPE != "" ? lang('Alter type') . ": " . h($TYPE) : lang('Create type')),
+	$error,
+	array(),
+	"",
+	($type_id === false)
+);
 
 if (!$row) {
 	$row["name"] = $TYPE;